% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdrf.R
\name{draw_glyph}
\alias{draw_glyph}
\title{draw_glyph}
\usage{
draw_glyph(fontfile, glyph)
}
\arguments{
\item{fontfile}{a raw vector representing a font file}

\item{glyph}{the character to be drawn. Can be text or an integer}
}
\value{
no return
}
\description{
Draws glyphs from a truetype font as grid grobs
}
\examples{
\dontrun{
if(interactive()){
 # ttf <- "raw vector with font file"
 draw_glyph(ttf, "a")
 }
}
}
