% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_obj_stream}
\alias{get_obj_stream}
\title{get_object_stream}
\usage{
get_obj_stream(filename, o)
}
\arguments{
\item{filename}{the path to a valid PDF file.}

\item{o}{The object number, given as an integer.}
}
\description{
Get a pdf object's stream (if any)
}
