% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdrf.R
\name{get_object}
\alias{get_object}
\title{Get the contents of a pdf object}
\usage{
get_object(pdf, number)
}
\arguments{
\item{pdf}{a valid pdf file location}

\item{number}{the object number}
}
\value{
a named vector of the dictionary and stream of the pdf object
}
\description{
Get the contents of a pdf object
}
\examples{
get_object(testfiles$leeds, 1)
}
