% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_xref}
\alias{get_xref}
\title{Return the xref table from a PDF document}
\usage{
get_xref(filename)
}
\arguments{
\item{filename}{the path to a valid PDF file.}
}
\description{
Returns a data frame of objects and their byte positions in a pdf
}
