% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdrf.R
\name{getglyphmap}
\alias{getglyphmap}
\title{Return map of glyphs from a page}
\usage{
getglyphmap(pdf, page = 1)
}
\arguments{
\item{pdf}{a valid pdf file location}

\item{page}{the page number from which to extract glyphs}
}
\value{
a dataframe of all entries of font encoding tables with width mapping
}
\description{
Used mainly for debugging, this function returns an R dataframe, one row for
each byte that may be used as a glyph. It shows the unicode number of
each interpreted glyph, as well as its width in text space.
}
\examples{
getglyphmap(pdfr_paths$leeds, 1)
}
