% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdrf.R
\name{getgrid}
\alias{getgrid}
\title{getgrid}
\usage{
getgrid(pdf, page)
}
\arguments{
\item{pdf}{a valid pdf file location}

\item{page}{the page number to be extracted}
}
\value{
a data frame of the page elements
}
\description{
Returns contents of a pdf page with 16x16 grid locations as bytes
}
\examples{
getgrid(testfiles$leeds, 1)
}
