% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdrf.R
\name{getpagestring}
\alias{getpagestring}
\title{pagestring}
\usage{
getpagestring(pdf, page)
}
\arguments{
\item{pdf}{a valid pdf file location}

\item{page}{the page number to be extracted}
}
\value{
a single string containing the page description program
}
\description{
Returns contents of a pdf page description program
}
\examples{
getpagestring(pdfr_paths$leeds, 1)
}
