% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfr.R
\name{hpdfr}
\alias{hpdfr}
\title{Print the head() of a data frame or tibble}
\usage{
hpdfr(x, nrows = 6)
}
\arguments{
\item{x}{A data frame or a tibble}

\item{nrows}{The number of rows to consider for head(). Default is 6}
}
\value{
The printed head of the data frame
}
\description{
Print the head() of a data frame or tibble
}
\examples{
hpdfr(iris)
iris \%>\% hpdfr()  ## when the dplyr package is loaded
iris \%>\% hpdfr(10)  ## print top 10 rows, when the dplyr package is loaded
}
