% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdrf.R
\name{internetFile}
\alias{internetFile}
\title{get a file from the internet}
\usage{
internetFile(x, filename = NULL)
}
\arguments{
\item{x}{a valid url}

\item{filename}{a name for the locally stored file (if required)}
}
\value{
either a character string or a raw vector
}
\description{
Returns the character string or raw vector (depending on content) from
a given url
}
\examples{
internetFile("http://www.google.com")
}
