% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdrf.R
\name{pdfboxes}
\alias{pdfboxes}
\title{pdfboxes}
\usage{
pdfboxes(pdf, pagenum)
}
\arguments{
\item{pdf}{a valid pdf file location}

\item{pagenum}{the page number to be plotted}
}
\value{
a ggplot
}
\description{
Plots the bounding boxes of text elements from a page as a ggplot.
}
\examples{
pdfboxes(pdfr_paths$leeds, 1)
}
