% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pdfdoc}
\alias{pdfdoc}
\title{pdfdoc}
\usage{
pdfdoc(filepath)
}
\arguments{
\item{filepath}{the path to a valid PDF file.}
}
\description{
List a PDF document's filename, catalogue dictionary, page dictionaries,
encryption status and file key (if encrypted)
}
