% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdrf.R
\name{pdfgrobs}
\alias{pdfgrobs}
\title{pdfgrobs}
\usage{
pdfgrobs(file_name, pagenum, scale = dev.size()[2]/10, enc = "UTF-8")
}
\arguments{
\item{file_name}{a valid pdf file location}

\item{pagenum}{the page number to be plotted}

\item{scale}{Document scale. Defaults to `dev.size()[2]/10`}

\item{enc}{Document encoding. Defaults to "UTF-8"}
}
\value{
invisibly returns grobs as well as drawing them
}
\description{
Plots the graphical elements of a pdf page as grobs
}
\examples{
pdfgrobs(testfiles$leeds, 1)
}
