% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pdfpageraw}
\alias{pdfpageraw}
\title{pdfpageraw}
\usage{
pdfpageraw(rawfile, pagenum)
}
\arguments{
\item{rawfile}{raw pdf data}

\item{pagenum}{the page to extract}
}
\description{
Returns a list comprising a page's fonts, its Postscript program as a text
string and its text elements as an R data frame
}
