% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdrf.R
\name{pdfplot}
\alias{pdfplot}
\title{pdfplot}
\usage{
pdfplot(pdf, page = 1, font = "Arial", textsize = 1)
}
\arguments{
\item{pdf}{a valid pdf file location}

\item{page}{the page number to be plotted}

\item{font}{the font to be used to show text on the plot}

\item{textsize}{the scale of the text to be shown}
}
\value{
a ggplot
}
\description{
Plots the text elements from a page as a ggplot.
The aim is not a complete pdf rendering but to help identify elements of
interest in the data frame of text elements to convert to data points.
}
\examples{
pdfplot(pdfref, 1)
}
