% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFR-package.R
\docType{package}
\name{PDFR-package}
\alias{PDFR}
\alias{PDFR-package}
\title{PDFR: Extract Text From PDFs In An R Friendly Way}
\description{
Extracts text from PDF into an R dataframe giving the content, size, position and font of any text elements. This information can then be manipulated in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AllanCameron/PDFR}
  \item Report bugs at \url{https://github.com/AllanCameron/PDFR/issues}
}

}
\author{
\strong{Maintainer}: Allan Cameron \email{Allan.Cameron@nhs.net} [copyright holder]

Other contributors:
\itemize{
  \item Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [contributor]
}

}
\keyword{internal}
