% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfr.R
\docType{package}
\name{pdfr}
\alias{pdfr}
\alias{pdfr-package}
\title{pdfr}
\usage{
pdfr(x)
}
\arguments{
\item{x}{A data frame or a tibble}
}
\value{
The printed data frame
}
\description{
Printing full data frames by typing `pdfr()` rather than typing `print.data.frame()`. This is useful for printing full data frames when piping tibbles through commands to ensure the entire data prints at the end
}
\examples{
pdfr(iris)
}
