% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rc4}
\alias{rc4}
\title{rc4}
\usage{
rc4(msg, key)
}
\arguments{
\item{msg}{The message to be hashed as a raw std::vector}

\item{key}{the raw std::vector with which to hash}
}
\description{
Performs an RC4 hash for a given key
}
