% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rex}
\alias{Rex}
\title{Rex}
\usage{
Rex(strvec, matcher)
}
\arguments{
\item{strvec}{The string(s) to be searched}

\item{matcher}{The regular expression to be matched}
}
\description{
Return substrings matching a given pattern
}
