\docType{data}
\name{run_testthat_tests}
\alias{run_testthat_tests}
\title{A tool used for symbol registration}
\format{A list of 4 fields
\describe{
  \item{name}{run_testthat_tests}
  \item{address}{a pointer to this symbol}
  \item{dll}{the compiled file where the symbol is contained}
  \item{numParameters}{no parameters}
}}
\usage{
run_testthat_tests
}
\description{
A registered native symbol used in testing
}
\keyword{tests}
