% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdrf.R
\name{segplot}
\alias{segplot}
\title{Show ggplot of page with segmentation lines}
\usage{
segplot(pdf, page = 1, atomic = FALSE, textsize = 1)
}
\arguments{
\item{pdf}{a valid pdf file location}

\item{page}{the page number from which to extract glyphs}

\item{atomic}{a boolean - should each letter treated individually?}

\item{textsize}{the size of the text to be shown on the plot}
}
\value{
no return - prints a ggplot
}
\description{
A way to assess the segmentation algorithm visually
}
\examples{
segplot(testfiles$leeds, 1)
}
