% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testfiles}
\alias{testfiles}
\title{Paths to test pdfs}
\format{
A list of 9 pdf files
\describe{
  \item{barcodes}{a pdf constructed in Rstudio}
  \item{chestpain}{a flow-chart for chest pain management}
  \item{pdfinfo}{information about the pdf format}
  \item{adobe}{an official adobe document}
  \item{leeds}{a table-rich local government document}
  \item{sams}{a document based on svg}
  \item{testreader}{a simple pdf test}
  \item{tex}{a simple tex test}
  \item{rcpp}{a CRAN package vignette}
}
}
\usage{
testfiles
}
\description{
A list of paths to locally stored test pdfs
}
\keyword{datasets}
