% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfr.R
\name{tpdfr}
\alias{tpdfr}
\title{Print the tail() of a data frame or tibble}
\usage{
tpdfr(x, nrows = 6)
}
\arguments{
\item{x}{A data frame or a tibble}

\item{nrows}{The number of rows to consider for tail(). Default is 6}
}
\value{
The printed tail of the data frame
}
\description{
Print the tail() of a data frame or tibble
}
\examples{
tpdfr(iris)
}
