% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_object}
\alias{get_object}
\title{get_object}
\usage{
get_object(filename, o)
}
\arguments{
\item{filename}{the path to a valid PDF file.}

\item{o}{The object number, given as an integer.}
}
\description{
Show the key:value pairs in a PDF object dictionary
}
