% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdrf.R
\name{get_object}
\alias{get_object}
\title{Get the contents of a pdf object}
\usage{
get_object(pdf, number)
}
\arguments{
\item{pdf}{a valid pdf file location}

\item{number}{the object number}
}
\value{
a named vector of the dictionary and stream of the pdf object
}
\description{
Returns a list consisting of a named vector representing key:value pairs
in a specified object. It also contains any stream data associated with
the object.
}
\examples{
get_object(pdfr_paths$leeds, 1)
}
