% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdrf.R
\name{get_xref}
\alias{get_xref}
\title{Get a pdf's xref table as an R dataframe}
\usage{
get_xref(pdf)
}
\arguments{
\item{pdf}{a valid pdf file location or raw data vector}
}
\value{
a data frame showing the positions of each object
}
\description{
Get a pdf's xref table as an R dataframe
}
\examples{
get_xref(testfiles$leeds)
}
