% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdrf.R
\name{pdfdoc}
\alias{pdfdoc}
\title{pdfdoc}
\usage{
pdfdoc(pdf)
}
\arguments{
\item{pdf}{a valid pdf file location}
}
\value{
a data frame of all text elements in a document
}
\description{
Returns contents of all pdf pages
}
\examples{
pdfdoc(testfiles$leeds)
}
