% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pdfpage}
\alias{pdfpage}
\title{pdfpage}
\usage{
pdfpage(filename, pagenum)
}
\arguments{
\item{filename}{the path to a valid PDF file.}

\item{pagenum}{the page to extract}
}
\description{
Returns a list comprising a page's fonts, its Postscript program as a text
string and its text elements as an R data frame
}
