% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdrf.R
\name{pdfpage}
\alias{pdfpage}
\title{pdfpage}
\usage{
pdfpage(pdf, page)
}
\arguments{
\item{pdf}{a valid pdf file location}

\item{page}{the page number to be plotted}
}
\value{
a list containing data frames
}
\description{
Returns contents of a pdf page
}
\examples{
pdfpage(testfiles$leeds, 1)
}
